c/Users/F12$/Desktop/restdelbot/requirements.txt
aiogram==3.4.1
anthropic==0.34.2
google-auth==2.28.1
google-api-python-client==2.119.0
python-dotenv==1.0.1